+++ src/main/java/redstonedubstep/mods/vanishmod/VanishEventListener.java
@@ -6,6 +6,7 @@ import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraftforge.event.ServerChatEvent;
+import net.minecraftforge.event.VanillaGameEvent;
 import net.minecraftforge.event.entity.PlaySoundAtEntityEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
@@ -28,6 +29,14 @@ public class VanishEventListener {
 
 			if (message instanceof TranslatableComponent component && component.getKey().contains("chat.type.text"))
 				event.setComponent(new TranslatableComponent("chat.type.text", new TextComponent("vanished").withStyle(ChatFormatting.GRAY), ((TranslatableComponent)message).getArgs()[1]));
+		}
+	}
+
+	@SubscribeEvent
+	public static void onVanillaGameEvent(VanillaGameEvent event) {
+		if (event.getCause() instanceof ServerPlayer player) {
+			if (VanishConfig.CONFIG.hidePlayersFromWorld.get() && VanishUtil.isVanished(player))
+				event.setCanceled(true);
 		}
 	}
 }